refactor(ocr): 迁移 OCR 初始化逻辑至 NTEScene 并优化相关任务参数
