+++ src/globals.py
@@ -15,43 +15,6 @@ class Globals(QObject):
         self.thread_pool_executor = None
         self.thread_pool_exit_event = Event()
         exit_event.bind_stop(self)
-        self._ocr_init = False
-        self.bg_ocr_lib = None
-        communicate.starting_emulator.connect(self.ocr_init)
-
-    def ocr_init(self, done, error, seconds_left):
-        if not self._ocr_init and done and error == "" and seconds_left == 0:
-            self._ocr_init = True
-            try:
-                all_tasks = og.executor.get_all_tasks()
-                if all_tasks and hasattr(all_tasks[0], "ocr"):
-                    logger.info("Warming up default OCR...")
-                    all_tasks[0].ocr(frame=np.zeros((50, 50, 3)))
-                
-                self.init_bg_ocr()
-                bg_ocr = getattr(og.executor, "_ocr_lib", {}).get("bg_onnx_ocr")
-                if bg_ocr:
-                    logger.info("Warming up background OCR...")
-                    bg_ocr.ocr(np.zeros((50, 50, 3), dtype=np.uint8))
-
-                logger.info("OCR initialization finished.")
-            except Exception as e:
-                logger.error(f"Failed to initialize OCR in background: {e}")
-
-    def init_bg_ocr(self):
-        from onnxocr.onnx_paddleocr import ONNXPaddleOcr
-
-        ocr_config = og.executor.config.get("ocr", {})
-        bg_config = ocr_config.get("bg_onnx_ocr") or ocr_config.get("default", {})
-        config_params = bg_config.get("params", {})
-
-        logger.info(f"Initializing bg onnxocr with params: {config_params}")
-        og.executor._ocr_lib["bg_onnx_ocr"] = ONNXPaddleOcr(
-            use_angle_cls=False,
-            logger=logger,
-            use_npu=config_params.get("use_npu", True),
-            use_openvino=config_params.get("use_openvino", False),
-        )
 
     def stop(self):
         self.shutdown_thread_pool_executor()
